feat: authorization using policy

+++ app/Http/Controllers/IdeaController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\IdeaRequest;
 use App\Models\Idea;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class IdeaController extends Controller
 {
@@ -18,11 +19,13 @@ class IdeaController extends Controller
 
     public function create()
     {
+        Gate::authorize('create', Idea::class);
         return view('ideas.create');
     }
 
     public function store(IdeaRequest $request)
     {
+        Gate::authorize('view', $idea);
         Auth::user()->ideas()->create([
             'description' => request('description'),
             'state' => 'pending',
@@ -32,6 +35,7 @@ class IdeaController extends Controller
 
     public function show(Idea $idea)
     {
+        Gate::authorize('view', $idea);
         return view('ideas.show', [
             'idea' => $idea
         ]);
@@ -39,6 +43,7 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        Gate::authorize('view', $idea);
         return view('ideas.edit', [
             'idea' => $idea
         ]);
@@ -46,6 +51,7 @@ class IdeaController extends Controller
 
     public function update(IdeaRequest $request, Idea $idea)
     {
+        Gate::authorize('view', $idea);
         $idea->update([
             'description' => request('description')
         ]);
@@ -54,6 +60,7 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+        Gate::authorize('view', $idea);
         $idea->delete();
         return redirect('/ideas');
